Design added

+++ Project/Controller/Salesman.php
@@ -9,7 +9,60 @@ class Controller_Salesman extends Controller_Admin_Action
             $this->redirect('login','admin_login');
         }
     }
-    
+
+    public function indexAction()
+    {
+        $this->setTitle('Salesman');
+        $content = $this->getLayout()->getContent();
+        $salesmanGrid = Ccc::getBlock('Salesman_Index');
+        $content->addChild($salesmanGrid);
+        $this->renderLayout();
+    }
+
+    public function gridBlockAction()
+    {
+        $salesmanGrid = Ccc::getBlock('Salesman_Grid')->toHtml();
+        $messageBlock = Ccc::getBlock('Core_Layout_Message')->toHtml();
+        $response = [
+            'status' => 'success',
+            'elements' => [
+                [
+                    'element' => '#indexContent',
+                    'content' => $salesmanGrid
+                ],
+                [
+                    'element' => '#adminMessage',
+                    'content' => $messageBlock
+                ]
+            ]
+        ];
+        $this->renderJson($response);
+    }
+
+    public function addBlockAction()
+    {
+        $salesmanModel = Ccc::getModel("Salesman");
+        Ccc::register('salesman',$salesmanModel);
+
+        $salesmanEdit = $this->getLayout()->getBlock('Salesman_Edit')->toHtml();
+        $messageBlock = Ccc::getBlock('Core_Layout_Message')->toHtml();
+        
+        $response = [
+            'status' => 'success',
+            'elements' => [
+                [
+                    'element' => '#indexContent',
+                    'content' => $salesmanEdit
+                ],
+                [
+                    'element' => '#adminMessage',
+                    'content' => $messageBlock
+                ]
+            ]
+        ];
+        $this->renderJson($response);
+    }
+
     public function gridAction()
     {
         $this->setTitle('Salesman');
@@ -60,13 +113,59 @@ class Controller_Salesman extends Controller_Admin_Action
                 throw new Exception("Unable to insert Salesman.");
             }
             $this->getMessage()->addMessage('Salesman Inserted succesfully.',1); 
-            $this->redirect('grid','salesman',[],true);
+            $this->gridBlockAction();
         }
         catch (Exception $e)
         {
             $this->getMessage()->addMessage($e->getMessage(),3);
-            $this->redirect('grid','Salesman',[],true);
+            $this->gridBlockAction();
         }
+    }
+
+    public function editBlockAction()
+    {
+        try
+        {
+            $this->setTitle('Edit Salesman');
+            $salesmanModel = Ccc::getModel('Salesman');
+            $request = $this->getRequest();
+            $id = (int)$request->getRequest('id');
+
+            if(!$id)
+            {
+                throw new Exception("Request Invalid.");
+            }
+            
+            $salesman = $salesmanModel->load($id);
+            
+            if(!$salesman)
+            {   
+                throw new Exception("System is unable to find record.");
+            }
+            Ccc::register('salesman',$salesman);
+
+            $salesmanEdit = Ccc::getBlock('Salesman_Edit')->toHtml();
+            $messageBlock = Ccc::getBlock('Core_Layout_Message')->toHtml();
+            $response = [
+                'status' => 'success',
+                'elements' => [
+                    [
+                        'element' => '#indexContent',
+                        'content' => $salesmanEdit
+                    ],
+                    [
+                        'element' => '#adminMessage',
+                        'content' => $messageBlock
+                    ]
+                ]
+            ];
+            $this->renderJson($response);
+        }
+        catch (Exception $e)
+        {
+            $this->getMessage()->addMessage($e->getMessage(),3);
+            $this->gridBlockAction();
+        }   
     }
 
     public function editAction()
@@ -120,12 +219,15 @@ class Controller_Salesman extends Controller_Admin_Action
             $salesmanId = (int)$request->getRequest('id');
 
             $customers = $customerModel->fetchAll("SELECT * FROM `customer` WHERE `salesmanId` = {$salesmanId}");
-            foreach($customers as $customer)
+            if($customers)
             {
-                $customerPrices = $customerPriceModel->fetchAll("SELECT `entityId` FROM `customer_price` WHERE `customerId` = {$customer->customerId}");
-                foreach ($customerPrices as $customerPrice) 
+                foreach($customers as $customer)
                 {
-                    $customerPriceModel->load($customerPrice->entityId)->delete();
+                    $customerPrices = $customerPriceModel->fetchAll("SELECT `entityId` FROM `customer_price` WHERE `customerId` = {$customer->customerId}");
+                    foreach ($customerPrices as $customerPrice) 
+                    {
+                        $customerPriceModel->load($customerPrice->entityId)->delete();
+                    }
                 }
             }
 
@@ -136,12 +238,12 @@ class Controller_Salesman extends Controller_Admin_Action
             }
             $result->delete();
             $this->getMessage()->addMessage('Data Deleted.');
-            $this->redirect('grid','salesman',[],true);
+            $this->gridBlockAction();
         } 
         catch (Exception $e)
         {
             $this->getMessage()->addMessage($e->getMessage(),3);
-            $this->redirect('grid','salesman',[],true);
+            $this->gridBlockAction();
         }
     }
 }
